requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
faiss-cpu==1.7.4
requests==2.31.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
httpx==0.26.0
